+++ MD/app/src/main/java/com/fimo/aidentist/ui/menu/treatment/TreatmentFragment.kt
@@ -47,6 +47,7 @@ class TreatmentFragment : Fragment() {
             val intent = Intent(activity, MainActivity::class.java)
             activity?.startActivity(intent)
         }
+        binding.progressBar.progress = 100
         return view
     }
 }
