+++ MD/app/src/main/res/drawable/background_button_accent_disco.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ripple xmlns:android="http://schemas.android.com/apk/res/android"
+    android:color="@color/black">
+    <item>
+        <selector>
+            <item android:state_enabled="false">
+                <shape
+                    android:shape="rectangle">
+                    <corners android:radius="10dp" />
+                    <solid android:color="@color/orange_yellow" />
+                </shape>
+            </item>
+
+            <item>
+                <shape
+                    android:shape="rectangle">
+                    <corners android:radius="10dp" />
+                    <solid android:color="@color/orange_peel" />
+                </shape>
+            </item>
+        </selector>
+    </item>
+</ripple>
