+++ MD/app/src/main/res/drawable/background_button_accent_health.xml
@@ -7,7 +7,7 @@
                 <shape
                     android:shape="rectangle">
                     <corners android:radius="10dp" />
-                    <solid android:color="@color/light_cornflower" />
+                    <solid android:color="@color/paris_green" />
                 </shape>
             </item>
 
@@ -15,7 +15,7 @@
                 <shape
                     android:shape="rectangle">
                     <corners android:radius="10dp" />
-                    <solid android:color="@color/blue_ivy" />
+                    <solid android:color="@color/paris_green" />
                 </shape>
             </item>
         </selector>
