+++ MD/app/src/main/res/drawable/ic_clean_teeth_hd.xml
@@ -1,0 +1,42 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="31dp"
+    android:height="28dp"
+    android:viewportWidth="31"
+    android:viewportHeight="28">
+  <path
+      android:pathData="M26.997,8.38C27.741,8.66 28.314,9.233 28.78,9.837C29.227,10.415 29.606,11.075 29.883,11.745C30.167,12.434 30.319,13.185 30.384,13.921C30.45,14.679 30.403,15.462 30.125,16.18C29.964,16.598 29.309,16.304 29.47,15.889C29.702,15.288 29.713,14.619 29.664,13.987C29.613,13.323 29.469,12.665 29.228,12.041C28.997,11.443 28.653,10.874 28.276,10.351C27.872,9.793 27.366,9.262 26.692,9.008C26.258,8.843 26.565,8.217 26.997,8.38Z"
+      android:fillColor="#231F20"/>
+  <path
+      android:pathData="M1.461,16.522C1.432,15.45 1.399,14.358 1.665,13.309C1.936,12.241 2.587,11.36 3.53,10.738C3.911,10.486 3.55,9.887 3.165,10.141C2.099,10.845 1.347,11.809 1.001,13.008C0.672,14.145 0.708,15.355 0.739,16.522C0.75,16.967 1.473,16.968 1.461,16.522Z"
+      android:fillColor="#231F20"/>
+  <path
+      android:pathData="M9.262,27.317C7.488,27.27 5.938,26.945 4.978,24.119C4.317,22.173 4.535,19.88 4.314,18.393C4.092,16.905 2.245,12.369 2.023,8.724C1.802,5.08 4.165,1.287 8.006,1.064C11.847,0.841 13.398,2.105 15.392,2.18C17.98,2.277 19.963,0.763 23.444,0.915C26.842,1.064 28.541,4.634 28.688,8.427C28.818,11.778 26.717,16.715 26.767,18.392C26.842,20.846 26.693,27.094 23.813,27.094C20.482,27.094 19.75,22.557 18.791,21.664C17.029,20.028 15.54,19.879 13.99,21.367C12.191,23.091 12.069,27.392 9.262,27.317Z"
+      android:fillColor="#DEECED"/>
+  <path
+      android:pathData="M10.159,2.862C8.818,2.528 7.704,2.91 6.792,3.636C6.02,4.057 5.304,4.826 4.985,5.366C4.23,6.641 3.715,8.379 4.485,9.778C5.127,10.943 6.659,10.939 7.149,9.674C7.582,8.557 7.194,7.362 7.946,6.317C8.666,5.316 9.693,5.314 10.698,4.681C11.806,3.984 11.194,3.12 10.159,2.862Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M28.691,8.431C28.556,4.983 27.14,1.719 24.33,1.042C29.456,3.051 27.842,10.908 25.409,14.156C24.735,15.057 23.698,15.24 22.398,16.115C21.197,16.923 24.18,16.99 25.12,19.594C25.681,21.147 26.372,24.99 24.084,25.653C23.335,25.871 22.296,25.906 21.604,25.607C21.16,25.416 20.544,24.916 20.194,24.313C20.924,25.706 21.972,27.098 23.817,27.098C26.698,27.098 26.846,20.851 26.772,18.396C26.72,16.72 28.822,11.783 28.691,8.431Z"
+      android:fillColor="#B9E1EA"/>
+  <path
+      android:pathData="M9.7,9.364C10.166,9.364 10.543,9.003 10.543,8.558C10.543,8.113 10.166,7.752 9.7,7.752C9.235,7.752 8.857,8.113 8.857,8.558C8.857,9.003 9.235,9.364 9.7,9.364Z"
+      android:fillColor="#231F20"/>
+  <path
+      android:pathData="M22.298,8.558C22.298,9.003 21.921,9.364 21.455,9.364C20.99,9.364 20.612,9.003 20.612,8.558C20.612,8.113 20.99,7.752 21.455,7.752C21.921,7.751 22.298,8.112 22.298,8.558Z"
+      android:fillColor="#231F20"/>
+  <path
+      android:pathData="M15.683,15.511C15.715,15.51 15.746,15.509 15.778,15.507C17.302,15.424 18.157,14.731 18.604,14.164C19.091,13.549 19.213,12.956 19.218,12.931C19.24,12.818 19.163,12.71 19.046,12.688C18.928,12.667 18.815,12.74 18.792,12.852C18.775,12.939 18.332,14.953 15.752,15.093C14.714,15.148 14.035,14.758 13.649,14.421C13.052,13.9 12.823,13.248 12.808,12.811C12.804,12.697 12.704,12.608 12.584,12.611C12.465,12.615 12.371,12.711 12.375,12.825C12.399,13.49 12.775,14.218 13.358,14.727C13.793,15.108 14.547,15.547 15.683,15.511Z"
+      android:fillColor="#231F20"/>
+  <path
+      android:pathData="M22.772,4.473C22.497,4.204 21.849,3.934 21.398,4.185C21.21,4.289 21.126,4.521 21.319,4.693C21.554,4.897 21.774,4.692 22.364,5.101C22.569,5.242 22.788,5.502 22.996,5.136C23.172,4.824 22.944,4.641 22.772,4.473Z"
+      android:fillColor="#231F20"/>
+  <path
+      android:pathData="M8.197,4.368C8.455,4.108 9.036,3.863 9.407,4.14C9.561,4.256 9.616,4.496 9.435,4.661C9.216,4.857 9.042,4.638 8.5,5.026C8.312,5.16 8.102,5.414 7.951,5.03C7.822,4.706 8.034,4.531 8.197,4.368Z"
+      android:fillColor="#231F20"/>
+  <path
+      android:pathData="M28.863,18.557C27.979,18.19 28.378,17.333 28.812,16.794C28.603,16.896 28.291,17.027 28.131,16.767C27.983,16.528 28.315,16.306 28.512,16.206C29.002,15.959 29.483,15.694 29.675,15.157L30.059,14.177C30.014,15.162 30.45,16.078 30.366,17.056C30.306,17.763 29.845,18.964 28.863,18.557Z"
+      android:fillColor="#231F20"/>
+  <path
+      android:pathData="M2.2,19.038C3.069,18.639 2.636,17.797 2.182,17.275C2.394,17.369 2.712,17.489 2.861,17.223C2.999,16.979 2.659,16.769 2.458,16.677C1.959,16.447 1.468,16.199 1.255,15.67L0.834,14.704C0.917,15.688 0.516,16.618 0.639,17.592C0.727,18.295 1.235,19.48 2.2,19.038Z"
+      android:fillColor="#231F20"/>
+</vector>
