+++ MD/app/src/main/res/layout/fragment_analisis.xml
@@ -9,7 +9,7 @@
 
 <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_height="wrap_content"
     tools:context=".ui.analisis.AnalisisFragment">
 
 
@@ -49,12 +49,12 @@
                 android:id="@+id/disease_title"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:layout_marginStart="8dp"
                 android:fontFamily="@font/nunito_bold"
                 android:text="@string/periodontal"
                 android:textColor="@color/black"
                 android:textSize="13sp"
-                app:layout_constraintStart_toEndOf="@+id/button"
+                app:layout_constraintEnd_toEndOf="@+id/desc_disease"
+                app:layout_constraintStart_toStartOf="@+id/desc_disease"
                 app:layout_constraintTop_toTopOf="@+id/button">
 
             </TextView>
@@ -64,28 +64,29 @@
                 android:layout_width="260dp"
                 android:layout_height="wrap_content"
                 android:layout_marginBottom="20dp"
-                android:paddingEnd="25dp"
-                android:paddingStart="5dp"
                 android:fontFamily="@font/nunito_regular"
                 android:justificationMode="inter_word"
+                android:paddingStart="5dp"
+                android:paddingEnd="25dp"
                 android:text="@string/periodontal_desc"
                 android:textColor="@color/black"
                 android:textSize="12sp"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="@+id/disease_title"
+                app:layout_constraintStart_toEndOf="@+id/button"
                 app:layout_constraintTop_toBottomOf="@+id/disease_title">
 
             </TextView>
 
-            <Button
+            <ImageView
+                android:contentDescription="@string/logo_periodental"
                 android:id="@+id/button"
-                android:layout_width="80dp"
-                android:layout_height="80dp"
+                android:layout_width="70dp"
+                android:layout_height="70dp"
                 android:layout_marginStart="24dp"
                 android:layout_marginTop="8dp"
-                android:background="@drawable/background_button_accent"
-                android:drawableTop="@drawable/ic_disease_teeth"
+                android:background="@drawable/background_button_accent_perio"
+                android:src="@drawable/ic_disease_teeth"
                 android:fontFamily="@font/nunito_medium"
                 android:padding="10dp"
                 android:text="@string/disease"
@@ -94,7 +95,7 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/dental_prediction">
 
-            </Button>
+            </ImageView>
 
         </androidx.constraintlayout.widget.ConstraintLayout>
