+++ MD/app/src/main/res/values/colors.xml
@@ -15,6 +15,7 @@
     <color name="red">#FF0000</color>
     <color name="pastel_magenta">#FA90C2</color>
     <color name="orange_peel">#FF9D00</color>
+    <color name="paris_green">#58CF99</color>
 
     <!--  Dark Theme Color  -->
     <color name="rangoon_green">#171717</color>
