+++ MD/app/src/main/res/values/strings.xml
@@ -168,6 +168,12 @@
     <string name="desc_disoloration">Diskolorasi ekstrinsik disebabkan karena adanya penodaan (staining) akibat warna bahan makanan dan minuman yang dikonsumsi. Penggunaan produk tembakau, teh, kopi, obat kumur tertentu, dan pigmen di dalam makanan menyebabkan terbentuknya noda yang menempel pada permukaan gigi maupun bahan tambal gigi.</string>
     <string name="risk_dicoloration">Diskolorasi pada gigi dapat diperbaiki dengan menggunakan bahan pemutih gigi atau yang disebut dengan bleaching. Bleaching adalah tindakan untuk meningkatkan warna gigi menjadi lebih putih dari sebelumnya. Jenisnya ada berbagai macam, antara lain menggunakan zat kimia hidrogen peroksida dan karbamid peroksida dengan berbagai konsentrasi.</string>
     <string name="scan_first">Scan gigi terlebih dahulu ya</string>
+    <string name="healthy_title">Healthy</string>
+    <string name="healthy_desc">Healthy Teeth, bagian atas dan bawah gigi tetap utuh, tidak ada lubang ataupun lekukan yang membuat kasar. Bebas plak, gigi yang putih dan tidak ada noda di sana-sini.</string>
+    <string name="risk_heath">Tentu tidak. Gigi anda termasuk dalam gigi sehat, untuk meningkatkanya, lakukan beberapa perawatan gigi seperti, gosok gigi secara teratur, dan menjaga makanan yang dimakan.</string>
+    <string name="logo_periodental">Logo Periodental</string>
+    <string name="logo_discoloration">Logo Discoloration</string>
+    <string name="logo_healthy">Logo Healhty</string>
 
     <!--  Database Local  -->
     <string-array name="name">
