+++ MD/app/src/main/res/values/styles.xml
@@ -46,7 +46,7 @@
         <item name="android:layout_height">50dp</item>
         <item name="fontFamily">@font/nunito_bold</item>
         <item name="android:textAllCaps">false</item>
-        <item name="android:background">@drawable/background_button_accent</item>
+        <item name="android:background">@drawable/background_button_accent_perio</item>
         <item name="android:textSize">16sp</item>
         <item name="android:textColor">@drawable/button_text_color_selector</item>
     </style>
